Flyttet funksjoner i riktige mapper
