+++ src/test/actions.tsx
@@ -1,0 +1,42 @@
+import { render } from './render';
+import { TestContainer } from './TestContainer';
+import { OvergangsstønadApp } from '../søknader/overgangsstønad/OvergangsstønadApp';
+import { Screen, waitFor, within } from '@testing-library/dom';
+import { expect } from 'vitest';
+import { UserEvent } from '@testing-library/user-event/index';
+
+export const navigerTilOmDeg = async () => {
+  const { screen, user } = render(
+    <TestContainer>
+      <OvergangsstønadApp />
+    </TestContainer>
+  );
+
+  await waitFor(() => {
+    expect(
+      screen.getByRole('heading', {
+        level: 1,
+        name: 'Søknad om overgangsstønad',
+      })
+    ).toBeInTheDocument();
+  });
+
+  await user.click(
+    screen.getByRole('button', {
+      name: 'Fortsett på søknaden',
+    })
+  );
+
+  return { screen, user };
+};
+
+export const klikkSvarRadioknapp = async (
+  groupName: string,
+  radioLabel: string,
+  screen: Screen,
+  user: UserEvent
+) => {
+  const radioGroup = screen.getByRole('group', { name: groupName });
+  const radio = within(radioGroup).getByRole('radio', { name: radioLabel });
+  await user.click(radio);
+};
